requirements.txt
dj-database-url==0.5.0
Django==2.1.15
django-appconf==1.0.4
django-compressor==1.4
django-crispy-forms==1.11.0
django-dynamic-formsets==0.0.8
django-heroku==0.3.1
django-pagination==1.0.7
django-photologue==2.3
django-solo==1.0.5
flake8>=3.6.0,<3.7.0
gunicorn==20.0.4
mccabe==0.6.1
Pillow==2.5.1
psycopg2==2.8.6
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2021.1
South==1.0
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2
